+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0 - 2020-08-08
+
+* ➕ Can access a google sheet and reads suggestions for the next event locations
+
 ## 1.0.0 - 2020-08-07
 
 * ➕ Add nomination of one group member to organize the next event.